pyproject.toml
[tool.poetry]
name = "Lumen"
version = "0.1.0"
description = "Lumen!"
authors = [""]

[tool.poetry.dependencies]
python = "^3.9"
aiohttp = "*"
dateparser = "*"
"discord.py" = {git = "https://github.com/Rapptz/discord.py.git", branch = "sharding-rework"}
pyyaml = "*"
aurcore = "*"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

